recurring task bar chart works fully, but now default is in minutes, so if user gives up after a few seconds it won't display anything

+++ app/src/main/java/sg/edu/np/mad/madassignment1/Task.java
@@ -128,7 +128,7 @@ public class Task {
         this.taskEndTime = taskEndTime;
     }
 
-    public double getTaskDuration() {
+    public long getTaskDuration() {
         return taskDuration;
     }
